scraper/Get_From_Cloud/requirements.txt
GetOldTweets3
bs4
requests
requests.auth
pillow
datetime